fix: Show sidebar on mobile when pressing the burger menu icon
